+++ modules/core/src/adapter/luma.ts
@@ -4,7 +4,8 @@
 
 import type {Log} from '@probe.gl/log';
 import type {DeviceProps} from './device';
-import {Device, DeviceFactory} from './device';
+import {Device} from './device';
+import {Adapter} from './adapter';
 import {StatsManager} from '../utils/stats-manager';
 import {lumaStats} from '../utils/stats-manager';
 import {log} from '../utils/log';
@@ -12,68 +13,62 @@ import {log} from '../utils/log';
 const ERROR_MESSAGE =
   'No matching device found. Ensure `@luma.gl/webgl` and/or `@luma.gl/webgpu` modules are imported.';
 
-const deviceMap = new Map<string, DeviceFactory>();
+const preregisteredAdapters = new Map<string, Adapter>();
 
 /** Properties for creating a new device */
 export type CreateDeviceProps = DeviceProps & {
   /** Selects the type of device. `best-available` uses webgpu if available, then webgl. */
   type?: 'webgl' | 'webgpu' | 'unknown' | 'best-available';
-  devices?: DeviceFactory[];
+  adapters?: Adapter[];
 };
 
 /** Properties for attaching an existing WebGL context or WebGPU device to a new luma Device */
 export type AttachDeviceProps = DeviceProps & {
   /** Externally created WebGL context or WebGPU device */
   handle: WebGL2RenderingContext; // | GPUDevice;
-  devices?: DeviceFactory[];
+  adapters?: Adapter[];
 };
 
 /**
  * Entry point to the luma.gl GPU abstraction
- * Register WebGPU and/or WebGL devices (controls application bundle size)
+ * Register WebGPU and/or WebGL adapters (controls application bundle size)
  * Run-time selection of the first available Device
  */
 export class luma {
   static defaultProps: Required<CreateDeviceProps> = {
     ...Device.defaultProps,
     type: 'best-available',
-    devices: undefined!
+    adapters: undefined!
   };
 
-  /** Global stats for all devices */
+  /** Global stats for all adapters */
   static stats: StatsManager = lumaStats;
 
   /** Global log */
   static log: Log = log;
 
-  static registerDevices(deviceClasses: DeviceFactory[]): void {
-    for (const deviceClass of deviceClasses) {
-      deviceMap.set(deviceClass.type, deviceClass);
+  static registerAdapters(adapters: Adapter[]): void {
+    for (const deviceClass of adapters) {
+      preregisteredAdapters.set(deviceClass.type, deviceClass);
     }
   }
 
-  static getAvailableDevices(): string[] {
-    // @ts-expect-error
-    return Array.from(deviceMap).map(Device => Device.type);
-  }
-
-  static getSupportedDevices(): string[] {
-    return (
-      Array.from(deviceMap)
-        // @ts-expect-error
-        .filter(Device => Device.isSupported())
-        // @ts-expect-error
-        .map(Device => Device.type)
-    );
+  /** Get type strings for supported Devices */
+  static getSupportedAdapters(adapters: Adapter[] = []): string[] {
+    const adapterMap = getAdapterMap(adapters);
+    return Array.from(adapterMap)
+      .map(([, adapter]) => adapter)
+      .filter(adapter => adapter.isSupported?.())
+      .map(adapter => adapter.type);
   }
 
   /** Get type strings for best available Device */
-  static getBestAvailableDeviceType(devices: DeviceFactory[] = []): 'webgpu' | 'webgl' | null {
-    const deviceMap = getDeviceMap(devices);
-    if (deviceMap.get('webgpu')?.isSupported?.()) {
+  static getBestAvailableAdapter(adapters: Adapter[] = []): 'webgpu' | 'webgl' | null {
+    const adapterMap = getAdapterMap(adapters);
+    if (adapterMap.get('webgpu')?.isSupported?.()) {
       return 'webgpu';
     }
-    if (deviceMap.get('webgl')?.isSupported?.()) {
+    if (adapterMap.get('webgl')?.isSupported?.()) {
       return 'webgl';
     }
     return null;
@@ -83,22 +78,46 @@ export class luma {
     Object.assign(luma.defaultProps, props);
   }
 
+  /** Creates a device. Asynchronously. */
+  static async createDevice(props: CreateDeviceProps = {}): Promise<Device> {
+    props = {...luma.defaultProps, ...props};
+
+    // Should be handled by attach device
+    // if (props.gl) {
+    //   props.type = 'webgl';
+    // }
+
+    const adapterMap = getAdapterMap(props.adapters);
+
+    let type: string = props.type || '';
+    if (type === 'best-available') {
+      type = luma.getBestAvailableAdapter(props.adapters) || type;
+    }
+
+    const adapters = getAdapterMap(props.adapters) || adapterMap;
+
+    const adapter = adapters.get(type);
+    const device = await adapter?.create?.(props);
+    if (device) {
+      return device;
+    }
+
+    throw new Error(ERROR_MESSAGE);
+  }
+
   /** Attach to an existing GPU API handle (WebGL2RenderingContext or GPUDevice). */
   static async attachDevice(props: AttachDeviceProps): Promise<Device> {
-    const devices = getDeviceMap(props.devices) || deviceMap;
+    const adapters = getAdapterMap(props.adapters);
 
     // WebGL
+    let type = '';
     if (props.handle instanceof WebGL2RenderingContext) {
-      const Device = devices.get('webgl');
-      const device = Device?.attach?.(null);
-      if (device) {
-        return device;
-      }
+      type = 'webgl';
     }
 
     // TODO - WebGPU does not yet have a stable API
     // if (props.handle instanceof GPUDevice) {
-    //   const WebGPUDevice = devices.get('webgpu') as any;
+    //   const WebGPUDevice = adapters.get('webgpu') as any;
     //   if (WebGPUDevice) {
     //     return (await WebGPUDevice.attach(props.handle)) as Device;
     //   }
@@ -106,32 +125,11 @@ export class luma {
 
     // null
     if (props.handle === null) {
-      const Device = devices.get('unknown');
-      const device = Device?.attach?.(null);
-      if (device) {
-        return device;
-      }
+      type = 'unknown';
     }
 
-    throw new Error(ERROR_MESSAGE);
-  }
-
-  /** Creates a device. Asynchronously. */
-  static async createDevice(props: CreateDeviceProps = {}): Promise<Device> {
-    props = {...luma.defaultProps, ...props};
-    if (props.gl) {
-      props.type = 'webgl';
-    }
-
-    const devices = getDeviceMap(props.devices) || deviceMap;
-
-    let type: string = props.type || '';
-    if (type === 'best-available') {
-      type = luma.getBestAvailableDeviceType(props.devices) || type;
-    }
-
-    const Device = devices.get(type);
-    const device = await Device?.create?.(props);
+    const adapter = adapters.get(type);
+    const device = adapter?.attach?.(null);
     if (device) {
       return device;
     }
@@ -160,18 +158,41 @@ export class luma {
     prototype.getContext = function (contextId: string, options?: WebGLContextAttributes) {
       // Attempt to force WebGL2 for all WebGL1 contexts
       if (contextId === 'webgl' || contextId === 'experimental-webgl') {
-        return this.originalGetContext('webgl2', options);
+        const context = this.originalGetContext('webgl2', options);
+        return context;
       }
       // For any other type, return the original context
       return this.originalGetContext(contextId, options);
     };
   }
+
+  /** Convert a list of adapters to a map */
+  protected getAdapterMap(adapters: Adapter[] = []): Map<string, Adapter> {
+    const map = new Map(preregisteredAdapters);
+    for (const adapter of adapters) {
+      map.set(adapter.type, adapter);
+    }
+    return map;
+  }
+
+  // DEPRECATED
+
+  /** @deprecated Use registerAdapters */
+  static registerDevices(deviceClasses: any[]): void {
+    log.warn('luma.registerDevices() is deprecated, use luma.registerAdapters() instead');
+    for (const deviceClass of deviceClasses) {
+      const adapter = deviceClass.adapter as Adapter;
+      if (adapter) {
+        preregisteredAdapters.set(adapter.type, adapter);
+      }
+    }
+  }
 }
 
-/** Convert a list of devices to a map */
-function getDeviceMap(deviceClasses: DeviceFactory[] = []): Map<string, DeviceFactory> {
-  const map = new Map<string, DeviceFactory>(deviceMap);
-  for (const deviceClass of deviceClasses) {
+/** Convert a list of adapters to a map */
+function getAdapterMap(adapters: Adapter[] = []): Map<string, Adapter> {
+  const map = new Map<string, Adapter>(preregisteredAdapters);
+  for (const deviceClass of adapters) {
     // assert(deviceClass.type && deviceClass.isSupported && deviceClass.create);
     map.set(deviceClass.type, deviceClass);
   }