+++ modules/webgpu/src/adapter/webgpu-device.ts
@@ -30,7 +30,7 @@ import type {
   QuerySet,
   QuerySetProps
 } from '@luma.gl/core';
-import {Device, DeviceFeatures, CanvasContext, log} from '@luma.gl/core';
+import {Device, DeviceFeatures} from '@luma.gl/core';
 import {WebGPUBuffer} from './resources/webgpu-buffer';
 import {WebGPUTexture} from './resources/webgpu-texture';
 import {WebGPUExternalTexture} from './resources/webgpu-external-texture';
@@ -49,8 +49,6 @@ import {WebGPUQuerySet} from './resources/webgpu-query-set';
 
 /** WebGPU Device implementation */
 export class WebGPUDevice extends Device {
-  static type: string = 'webgpu';
-
   /** type of this device */
   readonly type = 'webgpu';
 
@@ -72,78 +70,11 @@ export class WebGPUDevice extends Device {
   commandEncoder: GPUCommandEncoder | null = null;
   renderPass: WebGPURenderPass | null = null;
 
-  /** Check if WebGPU is available */
-  static isSupported(): boolean {
-    return Boolean(typeof navigator !== 'undefined' && navigator.gpu);
-  }
-
-  static async create(props: DeviceProps): Promise<WebGPUDevice> {
-    if (!navigator.gpu) {
-      throw new Error(
-        'WebGPU not available. Open in Chrome Canary and turn on chrome://flags/#enable-unsafe-webgpu'
-      );
-    }
-    log.groupCollapsed(1, 'WebGPUDevice created')();
-    const adapter = await navigator.gpu.requestAdapter({
-      powerPreference: 'high-performance'
-      // forceSoftware: false
-    });
-    if (!adapter) {
-      throw new Error('Failed to request WebGPU adapter');
-    }
-
-    const adapterInfo = await adapter.requestAdapterInfo();
-    log.probe(2, 'Adapter available', adapterInfo)();
-
-    const requiredFeatures: GPUFeatureName[] = [];
-    const requiredLimits: Record<string, number> = {};
-
-    if (props.requestMaxLimits) {
-      // Require all features
-      requiredFeatures.push(...(Array.from(adapter.features) as GPUFeatureName[]));
-
-      // Require all limits
-      // Filter out chrome specific keys (avoid crash)
-      const limits = Object.keys(adapter.limits).filter(
-        key => !['minSubgroupSize', 'maxSubgroupSize'].includes(key)
-      );
-      for (const key of limits) {
-        const limit = key as keyof GPUSupportedLimits;
-        const value = adapter.limits[limit];
-        if (typeof value === 'number') {
-          requiredLimits[limit] = value;
-        }
-      }
-    }
-
-    const gpuDevice = await adapter.requestDevice({
-      requiredFeatures,
-      requiredLimits
-    });
-
-    log.probe(1, 'GPUDevice available')();
-
-    if (typeof props.canvas === 'string') {
-      await CanvasContext.pageLoaded;
-      log.probe(1, 'DOM is loaded')();
-    }
-
-    const device = new WebGPUDevice(gpuDevice, adapter, adapterInfo, props);
-
-    log.probe(
-      1,
-      'Device created. For more info, set chrome://flags/#enable-webgpu-developer-features'
-    )();
-    log.table(1, device.info)();
-    log.groupEnd(1)();
-    return device;
-  }
-
   constructor(
+    props: DeviceProps,
     device: GPUDevice,
     adapter: GPUAdapter,
-    adapterInfo: GPUAdapterInfo,
-    props: DeviceProps
+    adapterInfo: GPUAdapterInfo
   ) {
     super({...props, id: props.id || 'webgpu-device'});
     this.handle = device;